requirements.txt
altgraph==0.17.4
colorama==0.4.6
imohash==1.1.0
llvmlite==0.44.0
mmh3==5.1.0
numba==0.61.0
numpy==2.1.3
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
psutil==7.0.0
pygame-ce==2.5.3
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pywin32-ctypes==0.2.3
tqdm==4.67.1
varint==1.0.2